pyproject.toml
[project]
name = "medium-crawler"
version = "0.1.0"
description = "Medium crawler with Gmail authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "greenlet>=3.2.0",
    "playwright>=1.56.0",
    "google-auth>=2.34.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.150.0",
    "python-dotenv>=1.0.1",
    "beautifulsoup4>=4.12.3",
]
